Split Dhall.Syntax
 - Moved functions and optics to Dhall.Syntax.Operations
 - Moved types to Dhall.Syntax.Types
 - Moved instances for the following classes to own modules in
   Dhall.Syntax.Instances:
    - Applicative
    - Bifunctor
    - Data
    - Eq
    - Foldable
    - Functor
    - Lift
    - Monad
    - Monoid
    - NFData
    - Ord
    - Pretty
    - Semigroup
    - Show
    - Traversable
   Only the following instances are in Dhall.Syntax.Types: Generic,
   IsString, Bounded, Enum
 - Monad Expr is using Monad Applicative and not the other way round
