+++ dhall/src/Dhall/Syntax/Instances/Foldable.hs
@@ -1,0 +1,16 @@
+{-# LANGUAGE DeriveFoldable     #-}
+{-# LANGUAGE StandaloneDeriving #-}
+
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Dhall.Syntax.Instances.Foldable () where
+
+import Dhall.Syntax.Types
+
+deriving instance Foldable (Binding s)
+deriving instance Foldable (Chunks s)
+deriving instance Foldable (PreferAnnotation s)
+deriving instance Foldable (RecordField s)
+deriving instance Foldable (FunctionBinding s)
+deriving instance Foldable FieldSelection
+deriving instance Foldable (Expr s)
