+++ dhall/src/Dhall/Syntax/Instances/Monoid.hs
@@ -1,0 +1,9 @@
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Dhall.Syntax.Instances.Monoid () where
+
+import Dhall.Syntax.Instances.Semigroup ()
+import Dhall.Syntax.Types
+
+instance Monoid (Chunks s a) where
+    mempty = Chunks [] mempty
