+++ dhall/src/Dhall/Syntax/Instances/Show.hs
@@ -1,0 +1,28 @@
+{-# LANGUAGE StandaloneDeriving #-}
+
+{-# OPTIONS_GHC -Wno-orphans #-}
+
+module Dhall.Syntax.Instances.Show () where
+
+import Dhall.Syntax.Types
+
+deriving instance Show Const
+deriving instance Show Var
+deriving instance (Show s, Show a) => Show (Binding s a)
+deriving instance Show DhallDouble
+deriving instance (Show s, Show a) => Show (Chunks s a)
+deriving instance (Show s, Show a) => Show (PreferAnnotation s a)
+deriving instance (Show s, Show a) => Show (RecordField s a)
+deriving instance (Show s, Show a) => Show (FunctionBinding s a)
+deriving instance Show s => Show (FieldSelection s)
+deriving instance Show WithComponent
+deriving instance (Show s, Show a) => Show (Expr s a)
+deriving instance Show Directory
+deriving instance Show File
+deriving instance Show FilePrefix
+deriving instance Show Scheme
+deriving instance Show URL
+deriving instance Show ImportType
+deriving instance Show ImportMode
+deriving instance Show ImportHashed
+deriving instance Show Import
