+++ dhall/src/Dhall/Syntax/Types.hs-boot
@@ -1,0 +1,7 @@
+module Dhall.Syntax.Types where
+
+data Var
+
+data Const
+
+data Expr s a
